+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+[unreleased]
+
+* Send an info message immediately after a user joins the community
+
+
 [2.0.3-alpha.1]
 
 * Temporarily hiding leave community button from Possible impersonation attack